+++ test/e2e/autoscaling/dns_autoscaling.go
@@ -30,6 +30,7 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2elog "k8s.io/kubernetes/test/e2e/framework/log"
+	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
 
 	"github.com/onsi/ginkgo"
@@ -102,7 +103,7 @@ var _ = SIGDescribe("DNS horizontal autoscaling", func() {
 	// This test is separated because it is slow and need to run serially.
 	// Will take around 5 minutes to run on a 4 nodes cluster.
 	ginkgo.It("[Serial] [Slow] kube-dns-autoscaler should scale kube-dns pods when cluster size changed", func() {
-		numNodes, err := framework.NumberOfRegisteredNodes(c)
+		numNodes, err := e2enode.TotalRegistered(c)
 		framework.ExpectNoError(err)
 
 		ginkgo.By("Replace the dns autoscaling parameters with testing parameters")
@@ -157,7 +158,7 @@ var _ = SIGDescribe("DNS horizontal autoscaling", func() {
 
 		ginkgo.By("Restoring cluster size")
 		setMigSizes(originalSizes)
-		err = framework.WaitForReadyNodes(c, numNodes, scaleDownTimeout)
+		err = e2enode.WaitForReadyNodes(c, numNodes, scaleDownTimeout)
 		framework.ExpectNoError(err)
 
 		ginkgo.By("Wait for kube-dns scaled to expected number")
