+++ test/e2e/framework/node/BUILD
@@ -3,20 +3,22 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
-        "framework.go",
-        "nodes.go",
+        "resource.go",
+        "wait.go",
     ],
-    importpath = "k8s.io/kubernetes/test/e2e/cloud",
+    importpath = "k8s.io/kubernetes/test/e2e/framework/node",
     visibility = ["//visibility:public"],
     deps = [
-        "//staging/src/k8s.io/apimachinery/pkg/api/errors:go_default_library",
+        "//pkg/controller/nodelifecycle:go_default_library",
+        "//pkg/util/system:go_default_library",
+        "//staging/src/k8s.io/api/core/v1:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/fields:go_default_library",
+        "//staging/src/k8s.io/apimachinery/pkg/util/wait:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
-        "//test/e2e/framework:go_default_library",
+        "//staging/src/k8s.io/client-go/rest:go_default_library",
         "//test/e2e/framework/log:go_default_library",
-        "//test/e2e/framework/node:go_default_library",
-        "//vendor/github.com/onsi/ginkgo:go_default_library",
-        "//vendor/github.com/onsi/gomega:go_default_library",
+        "//test/utils:go_default_library",
     ],
 )
