+++ test/e2e/scheduling/taint_based_evictions.go
@@ -27,6 +27,7 @@ import (
 	clientset "k8s.io/client-go/kubernetes"
 	schedulerapi "k8s.io/kubernetes/pkg/scheduler/api"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
 	e2epod "k8s.io/kubernetes/test/e2e/framework/pod"
 
 	"github.com/onsi/ginkgo"
@@ -123,9 +124,9 @@ var _ = SIGDescribe("TaintBasedEvictions [Serial]", func() {
 		node := nodeList.Items[0]
 
 		ginkgo.By(fmt.Sprintf("Blocking traffic from node %s to the master", nodeName))
-		host, err := framework.GetNodeExternalIP(&node)
+		host, err := e2enode.GetExternalIP(&node)
 		if err != nil {
-			host, err = framework.GetNodeInternalIP(&node)
+			host, err = e2enode.GetInternalIP(&node)
 		}
 		framework.ExpectNoError(err)
 		masterAddresses := framework.GetAllMasterAddresses(cs)
@@ -143,7 +144,7 @@ var _ = SIGDescribe("TaintBasedEvictions [Serial]", func() {
 			}
 
 			ginkgo.By(fmt.Sprintf("Expecting to see node %q becomes Ready", nodeName))
-			framework.WaitForNodeToBeReady(cs, nodeName, time.Minute*1)
+			e2enode.WaitForNodeToBeReady(cs, nodeName, time.Minute*1)
 			ginkgo.By("Expecting to see unreachable=:NoExecute taint is taken off")
 			err := framework.WaitForNodeHasTaintOrNot(cs, nodeName, taint, false, time.Second*30)
 			framework.ExpectNoError(err)
@@ -154,7 +155,7 @@ var _ = SIGDescribe("TaintBasedEvictions [Serial]", func() {
 		}
 
 		ginkgo.By(fmt.Sprintf("Expecting to see node %q becomes NotReady", nodeName))
-		if !framework.WaitForNodeToBeNotReady(cs, nodeName, time.Minute*3) {
+		if !e2enode.WaitForNodeToBeNotReady(cs, nodeName, time.Minute*3) {
 			framework.Failf("node %q doesn't turn to NotReady after 3 minutes", nodeName)
 		}
 		ginkgo.By("Expecting to see unreachable=:NoExecute taint is applied")
