+++ test/e2e/storage/flexvolume.go
@@ -29,6 +29,7 @@ import (
 	apierrs "k8s.io/apimachinery/pkg/api/errors"
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/kubernetes/test/e2e/framework"
+	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
 	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
 	"k8s.io/kubernetes/test/e2e/framework/testfiles"
 	"k8s.io/kubernetes/test/e2e/framework/volume"
@@ -77,9 +78,9 @@ func installFlex(c clientset.Interface, node *v1.Node, vendor, driver, filePath 
 	host := ""
 	var err error
 	if node != nil {
-		host, err = framework.GetNodeExternalIP(node)
+		host, err = e2enode.GetExternalIP(node)
 		if err != nil {
-			host, err = framework.GetNodeInternalIP(node)
+			host, err = e2enode.GetInternalIP(node)
 		}
 	} else {
 		masterHostWithPort := framework.GetMasterHost()
@@ -106,9 +107,9 @@ func uninstallFlex(c clientset.Interface, node *v1.Node, vendor, driver string) 
 	host := ""
 	var err error
 	if node != nil {
-		host, err = framework.GetNodeExternalIP(node)
+		host, err = e2enode.GetExternalIP(node)
 		if err != nil {
-			host, err = framework.GetNodeInternalIP(node)
+			host, err = e2enode.GetInternalIP(node)
 		}
 	} else {
 		masterHostWithPort := framework.GetMasterHost()
